+++ test/dispatch_source_watcher_test.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:dispatch_source_watcher/dispatch_source_watcher.dart';
+import 'package:dispatch_source_watcher/src/dispatch_source_event.dart';
+import 'package:dispatch_source_watcher/src/dispatch_source_callback_aggregator.dart';
 import 'package:dispatch_source_watcher/src/dispatch_source_watcher_method_channel.dart';
+import 'package:dispatch_source_watcher/src/dispatch_source_watcher_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 class MockDispatchSourceWatcherPlatform 
@@ -19,6 +21,11 @@ class MockDispatchSourceWatcherPlatform
     return Future.value();
   }
 
+  @override
+  Future<void> stopWatching(String path) {
+    return Future.value();
+  }
+
   void triggerMockEvent(Map<String, Object> event) {
     _streamController.add(event);
   }
@@ -32,43 +39,43 @@ void main() {
   });
 
   test('watch and unwatch a file', () async {
-    final watcher = DispatchSourceWatcher();
+    final watcher = DispatchSourceCallbackAggregator();
     MockDispatchSourceWatcherPlatform fakePlatform = MockDispatchSourceWatcherPlatform();
     DispatchSourceWatcherPlatform.instance = fakePlatform;
     void callback1(event){}
-    watcher.watch('/dir/file1', callback1);
+    watcher.addCallback('/dir/file1', callback1);
     expect(watcher.callbacksForPath('/dir/file1'), equals([callback1]));
 
-    watcher.cancelWatchCallback('/dir/file1', callback1);
+    watcher.removeCallback('/dir/file1', callback1);
     expect(watcher.callbacksForPath('/dir/file1'), equals([]));
   });
 
   test('watch and unwatch a file multiple times', () async {
-    final watcher = DispatchSourceWatcher();
+    final watcher = DispatchSourceCallbackAggregator();
     MockDispatchSourceWatcherPlatform fakePlatform = MockDispatchSourceWatcherPlatform();
     DispatchSourceWatcherPlatform.instance = fakePlatform;
     void callback1(event){}
     void callback2(event){}
 
-    watcher.watch('/dir/file1', callback1);
-    watcher.watch('/dir/file1', callback2);
+    watcher.addCallback('/dir/file1', callback1);
+    watcher.addCallback('/dir/file1', callback2);
     expect(watcher.callbacksForPath('/dir/file1'), equals([callback1, callback2]));
 
-    watcher.cancelWatchCallback('/dir/file1', callback1);
+    watcher.removeCallback('/dir/file1', callback1);
     expect(watcher.callbacksForPath('/dir/file1'), equals([callback2]));
   });
 
   test('callback triggers on event', () async {
-    final watcher = DispatchSourceWatcher();
+    final watcher = DispatchSourceCallbackAggregator();
     MockDispatchSourceWatcherPlatform fakePlatform = MockDispatchSourceWatcherPlatform();
     DispatchSourceWatcherPlatform.instance = fakePlatform;
-    watcher.initialize();
+    watcher.ensureInitialized();
     final triggeredEvents = <Map<String, Object>>[];
     callback1(DispatchSourceEvent event){
       triggeredEvents.add(event.asDictionary());
     }
 
-    watcher.watch('/dir/file1', callback1);
+    watcher.addCallback('/dir/file1', callback1);
     expect(triggeredEvents, equals([]));
 
     final mockEvent1 = {"path":'/dir/file1', "event":2, "eventNames":["write"]};