+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+* Changed DispatchSourceWatcher to provide a stream (rather than callbacks)
+* Fixed low-level DispatchSource cancellation
+
 ## 0.1.0
 
 * first release